Refatorando valorinicial de double para decimal

+++ CalculaJuros.Domain/Extensions/TruncateExtension.cs
@@ -4,9 +4,9 @@ namespace CalculaJuros.Domain.Extensions
 {
     public static class TruncateExtension
     {
-        public static double Truncate(this double num, int precision)
+        public static decimal Truncate(this decimal num, int precision)
         {
-            double y = Math.Pow(10, precision);
+            decimal y = (decimal)Math.Pow(10, precision);
             return Math.Truncate(num * y) / y;
         }
     }
